requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.2
django-admin-list-filter-dropdown==1.0.2
django-ckeditor==5.4.0
django-classy-tags==0.8.0
django-debugtools==1.7.3
django-enumfield==1.2.1
django-formtools==2.1
django-import-export==1.1.0
django-js-asset==1.0.0
django-json-widget==0.1.1
django-mptt==0.9.1
django-mptt-admin==0.6.0
django-nested-admin==3.0.21
django-nested-inline==0.3.7
django-sekizai==0.10.0
django-treebeard==4.2.1
djangocms-admin-style==1.2.7
docutils==0.14
et-xmlfile==1.0.1
ExifRead==2.1.2
Flask==0.12.2
future==0.16.0
gunicorn==19.8.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isodate==0.6.0
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
msgpack-python==0.5.6
mysqlclient==1.3.12
natsort==5.3.3
oauthlib==2.0.7
odfpy==1.3.6
openpyxl==2.5.8
pkgconfig==1.3.1
<<<<<<< HEAD
<<<<<<< HEAD
psycopg2==2.7.3.2
=======
psycopg2>=2.7.3.2
>>>>>>> c36584180672367013b4df628472a56e547f5441
pycparser==2.19
=======
psycopg2>=2.7.3.2
>>>>>>> testing phase1
pycryptodomex==3.6.1
PyHamcrest==1.9.0
pyjwkest==1.4.0
PyJWT==1.6.1
pymarc==3.1.10
pyodbc==4.0.21
pyparsing==2.2.0
pypyodbc==1.3.4
python-dateutil==2.6.1
python-monkey-business==1.0.0
python3-openid==3.1.0
# NOTE: python3-saml installs xmlsec which needs special install on windows
# so comment this out and xmlsec below for windows
python3-saml==1.4.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
regex==2018.2.8
requests==2.19.1
requests-oauthlib==0.8.0
schedule==0.5.0
six==1.11.0
social==0.1.dev3
social-auth-app-django==2.1.0
social-auth-core==1.7.0
SQLAlchemy==1.2.7
tablib==0.12.1
txaio==18.8.1
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
xmlsec==1.3.52.dev0
zope.interface==4.5.0
